improvement of Slider, undo changes in SequenceSimulation

+++ src/GameMain/Gamemaster.cc
@@ -11,8 +11,7 @@ using namespace std;
 
 Gamemaster::Gamemaster(){};
 
-void Gamemaster::makeGame(unsigned int k, unsigned int length, unsigned int number_of_sequences,
-                          unsigned int probability){
+void Gamemaster::makeGame(unsigned int k, unsigned int length, unsigned int number_of_sequences, double probability){
     simulate(number_of_sequences, length, probability);
     GameGraph.resetGraph();
     GameGraph.readFastaFiles("sequences.fa", k);
